refactor: loaded-survey 추가 및 훅 이전
